requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
boto3==1.9.163
botocore==1.12.163
certifi==2019.6.16
chardet==3.0.4
dataclasses==0.6
discord-webhook==0.4.1
discord.py==1.2.3
dj-database-url==0.5.0
Django==2.2.1
django-background-tasks==1.2.0
django-ckeditor==5.7.1
django-cleanup==3.2.0
django-compat==1.0.15
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.7.1
django-widget-tweaks==1.4.3
djongo==1.2.32
dnspython==1.16.0
docutils==0.14
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
multidict==4.5.2
natsort==6.0.0
Pillow==6.0.0
psycopg2==2.8.2
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.2.4
urllib3==1.25.3
websockets==6.0
whitenoise==4.1.2
yarl==1.3.0